unificando logica de control de acceso a datos

+++ protected/controllers/OrganizacionController.php
@@ -123,12 +123,22 @@ class OrganizacionController extends Controller
     public function actionDelete($id)
     {
         if (Yii::app()->request->isPostRequest) {
-// we only allow deletion via POST request
-            $this->loadModel($id)->delete();
-
-// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
-            if (!isset($_GET['ajax']))
+            try{
+                $area = Area::model()->findByAttributes(['organizacion_id'=>$id]);
+                if(!is_null($area)){
+                    throw new Exception("Error. Esta organizacion tiene un area relacionada.");
+                }
+                $this->loadModel($id)->delete();
+                $data = "Se elimino correctamente la organizacion";
+                echo CJSON::encode($data);
+            }catch (Exception $exception){
+                $data = $exception->getMessage();
+                echo CJSON::encode($data);
+                die();
+            }
+            if (!isset($_GET['ajax'])) {
                 $this->redirect(isset($_POST['returnUrl']) ? $_POST['returnUrl'] : array('admin'));
+            }
         } else
             throw new CHttpException(400, 'Invalid request. Please do not repeat this request again.');
     }
